restructure test
